Move LimitedStringWriter into reporting utility module
The reporting utility imported LimitedStringWriter from the API handler module
(reports.rs), inverting the dependency direction for reusable rendering code.
Move the writer (and its unit tests) next to its only user in
src/utilities/reporting.rs so the utility no longer depends on a handler.
